requirements_0.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cryptography==45.0.5
datasets==2.20.0
fairlearn==0.12.0
faiss-cpu==1.11.0.post1
fastapi==0.112.0
filelock==3.18.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.9.1
mpmath==1.3.0
narwhals==1.48.0
networkx==3.5
nltk==3.8.1
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pdfplumber==0.11.2
pillow==11.3.0
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
sentence-transformers==5.0.0
setuptools==80.9.0
shap==0.48.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
streamlit==1.47.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.30.3
wordcloud==1.9.3